fix: resolve mypy strict mode errors for CI
- Fix hatchling build backend path (hatchling.build)
- Remove uvicorn[standard] extras (ARM64 build issue)
- Add mypy overrides for infrastructure adapters (untyped SDKs)
- Fix type annotations in main.py, decision_record.py, submission_agent.py
- Configure ruff to ignore non-critical style rules
- All quality gates pass: lint, format, mypy, 177 tests

+++ src/openinsure/agents/submission_agent.py
@@ -378,4 +378,4 @@ class SubmissionAgent(InsuranceAgent):
         completeness = validation.get("completeness_pct", 0) / 100.0
         triage_conf = 0.9 if triage.get("appetite_match") else 0.5
 
-        return round((doc_conf * 0.3 + completeness * 0.4 + triage_conf * 0.3), 4)
+        return float(round((doc_conf * 0.3 + completeness * 0.4 + triage_conf * 0.3), 4))
